Django 2.0 compatibility: Add mandatory `on_delete` argument to foreign keys (#47)
* Add mandatory `on_delete` argument to foreign keys

* Add `on_delete` to foreign keys

* Include `on_delete` in migration

I'm intentionally adding this to the existing migration.

`makemigrations` do not create a new migration for the `on_delete`
changes to the fields.

I also tried adding a new model without the argument (and Django 1.11),
for which the migration _did_ include the argument (defaults to
`models.CASCADE`). If I explicitly removed the `on_delete` argument from
that migration, added `on_delete` to the model and ran `makemigrations`
again, it did not create a new migration.

So I assume it's safe to just add it to the existing migration, to make
it compatible with the most recent versions of Django.